Add Timer and Pause Menu
-Add Time into the game, stay at Top Middle
-Add Pause Logic when press 'esc' on keyboard, it will show the pause menu, which will show current HP and timer, press RESUME will continue the game, press MAIN MENU will go back to starting page.

Note 1: The pause system currently uses instance_deactivate_all(), so the game world is not visible when paused, hides the game (you can't see the background).
Note 2: The pause logic is in the "objHPBarTimerPause", which will be under "objUI", and that's right it have 3 functions in one object right now, idk if it's a good thing or bad thing, might need to do something about it.
Note 3: I might have broken something in the game with this update, so it'd be good if you guys check the functionality of the game to see if it's still working like you expected it to work.